+++ postgres/src/driver.rs
@@ -184,7 +184,7 @@ async fn dns_resolve<'p>(host: &'p str, ports: &'p [u16]) -> Result<impl Iterato
 /// drop(drv);
 ///
 /// // client will always return error when it's driver is gone.
-/// let e = cli.query_simple("SELECT 1").unwrap_err();
+/// let e = cli.query("SELECT 1").unwrap_err();
 /// // a shortcut method can be used to determine if the error is caused by a shutdown driver.
 /// assert!(e.is_driver_down());
 ///