+++ postgres/src/lib.rs
@@ -108,7 +108,7 @@ pub mod dev {
 
     pub use crate::client::ClientBorrowMut;
     pub use crate::copy::r#Copy;
-    pub use crate::driver::codec::{AsParams, Encode, Response};
+    pub use crate::driver::codec::{Encode, Response};
     pub use crate::prepare::Prepare;
     pub use crate::query::Query;
 }
@@ -119,7 +119,9 @@ use xitca_io::io::AsyncIo;
 
 static NEXT_ID: AtomicUsize = AtomicUsize::new(0);
 
-type ZeroParam = [i32; 0];
+const fn zero_params() -> [i32; 0] {
+    []
+}
 
 #[derive(Debug)]
 pub struct Postgres {