+++ postgres/src/query.rs
@@ -12,10 +12,8 @@ use core::{
 };
 
 use super::{
-    driver::codec::{AsParams, Encode, IntoStream, Response},
+    driver::codec::{Encode, IntoStream, Response},
     error::Error,
-    iter::slice_iter,
-    types::ToSql,
 };
 
 /// trait generic over api used for querying with typed prepared statement.
@@ -39,7 +37,7 @@ pub trait Query {
     /// // prepare a statement with client type.
     /// let stmt = client._prepare("SELECT id from users", &[Type::INT4]).await?;
     /// // query with statement and typed params for a stream of rows
-    /// let mut stream = client._query(&stmt, &[&996i32])?;
+    /// let mut stream = client._query(stmt.bind([&996i32]))?;
     /// // obtain the first row and get user id.
     /// let row = stream.try_next().await?.unwrap();      
     /// let _id: i32 = row.try_get("id")?;
@@ -59,7 +57,7 @@ pub trait Query {
     /// let stmt = Statement::unnamed(client, "SELECT id from users", &[Type::INT4]);
     /// // query with the unnamed statement.
     /// // under the hood the statement is prepared in background and used for query and stream row parsing
-    /// let mut stream = client._query(stmt, &[&996i32])?;
+    /// let mut stream = client._query((stmt, [&996i32]))?;
     /// // obtain the first row and get user id.
     /// let row = stream.try_next().await?.unwrap();      
     /// let _id: i32 = row.try_get("id")?;
@@ -74,7 +72,7 @@ pub trait Query {
     /// # use xitca_postgres::{dev::{Prepare, Query}, statement::Statement, types::Type, AsyncLendingIterator, Client, Error};
     /// # async fn simple_query(client: &Client) -> Result<(), Error> {
     /// // query with a string. the string can contain multiple sql query and they have to be separated by semicolons
-    /// let mut stream = client._query("SELECT 1;SELECT 1", &[])?;
+    /// let mut stream = client._query("SELECT 1;SELECT 1")?;
     /// let _ = stream.try_next().await?;      
     /// # Ok(())
     /// # }
@@ -82,54 +80,30 @@ pub trait Query {
     /// [`Statement`]: crate::statement::Statement
     /// [`StatementUnnamed`]: crate::statement::StatementUnnamed
     #[inline]
-    fn _query<'a, S>(
-        &self,
-        stmt: S,
-        params: &[&(dyn ToSql + Sync)],
-    ) -> Result<<S::Output<'a> as IntoStream>::RowStream<'a>, Error>
+    fn _query<'a, S>(&self, stmt: S) -> Result<<S::Output<'a> as IntoStream>::RowStream<'a>, Error>
     where
         S: Encode + 'a,
     {
-        self._query_raw(stmt, slice_iter(params))
-    }
-
-    /// flexible version of [Query::_query]
-    #[inline]
-    fn _query_raw<'a, S, I>(&self, stmt: S, params: I) -> Result<<S::Output<'a> as IntoStream>::RowStream<'a>, Error>
-    where
-        S: Encode + 'a,
-        I: AsParams,
-    {
-        self._send_encode_query(stmt, params)
+        self._send_encode_query(stmt)
             .map(|(stream, res)| stream.into_stream(res))
     }
 
     /// query that don't return any row but number of rows affected by it
     #[inline]
-    fn _execute<S>(&self, stmt: S, params: &[&(dyn ToSql + Sync)]) -> ExecuteFuture
+    fn _execute<S>(&self, stmt: S) -> ExecuteFuture
     where
         S: Encode,
     {
-        self._execute_raw(stmt, slice_iter(params))
-    }
-
-    /// flexible version of [Query::_execute]
-    fn _execute_raw<S, I>(&self, stmt: S, params: I) -> ExecuteFuture
-    where
-        S: Encode,
-        I: AsParams,
-    {
-        let res = self._send_encode_query(stmt, params).map(|(_, res)| res).map_err(Some);
+        let res = self._send_encode_query(stmt).map(|(_, res)| res).map_err(Some);
         // TODO:
         // use async { res?.try_into_row_affected().await } with Rust 2024 edition
         ExecuteFuture { res, rows_affected: 0 }
     }
 
     /// encode statement and params and send it to client driver
-    fn _send_encode_query<'a, S, I>(&self, stmt: S, params: I) -> Result<(S::Output<'a>, Response), Error>
+    fn _send_encode_query<'a, S>(&self, stmt: S) -> Result<(S::Output<'a>, Response), Error>
     where
-        S: Encode + 'a,
-        I: AsParams;
+        S: Encode + 'a;
 }
 
 pub struct ExecuteFuture {