+++ postgres/src/transaction/portal.rs
@@ -3,7 +3,7 @@ use core::sync::atomic::Ordering;
 use postgres_protocol::message::backend;
 
 use crate::{
-    driver::codec::{AsParams, Encode, IntoStream, PortalCancel, PortalCreate, PortalQuery},
+    driver::codec::{Encode, IntoParams, IntoStream, PortalCancel, PortalCreate, PortalQuery},
     error::Error,
     query::Query,
     statement::Statement,
@@ -27,9 +27,7 @@ where
     C: Query,
 {
     fn drop(&mut self) {
-        let _ = self
-            .cli
-            ._send_encode_query::<_, crate::ZeroParam>(PortalCancel { name: &self.name }, []);
+        let _ = self.cli._send_encode_query(PortalCancel { name: &self.name });
     }
 }
 
@@ -39,18 +37,18 @@ where
 {
     pub(crate) async fn new<'p, I>(cli: &'p C, stmt: &'p Statement, params: I) -> Result<Portal<'p, C>, Error>
     where
-        I: AsParams,
+        I: IntoParams,
     {
         let name = format!("p{}", crate::NEXT_ID.fetch_add(1, Ordering::Relaxed));
 
-        let (_, mut res) = cli._send_encode_query(
+        let (_, mut res) = cli._send_encode_query((
             PortalCreate {
                 name: &name,
                 stmt: stmt.name(),
                 types: stmt.params(),
             },
             params,
-        )?;
+        ))?;
 
         match res.recv().await? {
             backend::Message::BindComplete => {}
@@ -64,13 +62,10 @@ where
         &self,
         max_rows: i32,
     ) -> Result<<<PortalQuery<'_> as Encode>::Output<'_> as IntoStream>::RowStream<'_>, Error> {
-        self.cli._query_raw::<_, crate::ZeroParam>(
-            PortalQuery {
-                name: &self.name,
-                columns: self.stmt.columns(),
-                max_rows,
-            },
-            [],
-        )
+        self.cli._query(PortalQuery {
+            name: &self.name,
+            columns: self.stmt.columns(),
+            max_rows,
+        })
     }
 }